requirements.txt
Django>=5.0,<6.0
Pillow
django-crispy-forms
crispy-bootstrap5
python-dotenv
google-generativeai
PyPDF2
